FINALAGENTATHON/requirements.txt
# Core FastAPI dependencies
fastapi>=0.104.0
uvicorn[standard]>=0.24.0
pydantic>=2.0.0
python-multipart>=0.0.6

# Google AI dependencies
google-genai>=1.56.0
google-generativeai>=0.8.6

# HTTP and networking
httpx>=0.25.0
requests>=2.31.0

# Data processing and validation
jsonschema>=4.20.0
json5>=0.9.14

# Job scheduling and background tasks
apscheduler>=3.10.0

# AI and machine learning (optional)
beeai>=0.1.0

# PDF generation (optional)
reportlab>=4.0.0

# Environment management
python-dotenv>=1.0.0

# Development and testing
pytest>=7.4.0
pytest-asyncio>=0.21.0

# Type hints and validation
typing-extensions>=4.8.0

# Date and time utilities
python-dateutil>=2.8.0

# Logging
structlog>=23.2.0